refactor: extract class `TerminalViewPreferencesDataStore.java`
The result helps maintain adherence to the Single Responsibility Principle. The code of your classes will be more obvious and understandable.
