+++ app/src/main/java/com/termux/app/fragments/settings/termux/TerminalViewPreferencesFragment.java
@@ -9,6 +9,7 @@ import androidx.preference.PreferenceFragmentCompat;
 import androidx.preference.PreferenceManager;
 
 import com.termux.R;
+import com.termux.app.datastore.termux.TerminalViewPreferencesDataStore;
 import com.termux.shared.termux.settings.preferences.TermuxAppSharedPreferences;
 
 @Keep
@@ -27,51 +28,3 @@ public class TerminalViewPreferencesFragment extends PreferenceFragmentCompat {
 
 }
 
-class TerminalViewPreferencesDataStore extends PreferenceDataStore {
-
-    private final Context mContext;
-    private final TermuxAppSharedPreferences mPreferences;
-
-    private static TerminalViewPreferencesDataStore mInstance;
-
-    private TerminalViewPreferencesDataStore(Context context) {
-        mContext = context;
-        mPreferences = TermuxAppSharedPreferences.build(context, true);
-    }
-
-    public static synchronized TerminalViewPreferencesDataStore getInstance(Context context) {
-        if (mInstance == null) {
-            mInstance = new TerminalViewPreferencesDataStore(context);
-        }
-        return mInstance;
-    }
-
-
-
-    @Override
-    public void putBoolean(String key, boolean value) {
-        if (mPreferences == null) return;
-        if (key == null) return;
-
-        switch (key) {
-            case "terminal_margin_adjustment":
-                    mPreferences.setTerminalMarginAdjustment(value);
-                break;
-            default:
-                break;
-        }
-    }
-
-    @Override
-    public boolean getBoolean(String key, boolean defValue) {
-        if (mPreferences == null) return false;
-
-        switch (key) {
-            case "terminal_margin_adjustment":
-                return mPreferences.isTerminalMarginAdjustmentEnabled();
-            default:
-                return false;
-        }
-    }
-
-}
